% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{adj2formula}
\alias{adj2formula}
\title{Form pair-wise formula from adjacency matrix.
XXXX}
\usage{
adj2formula(adj.mat, Xoption = FALSE)
}
\arguments{
\item{edge.mat}{The XX}
}
\value{
The function will XX
}
\description{
Assumes edges/nodes are numeric. Uses row and col numbers as node names.
}
\details{
The function will XXXX
}
