% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_util.R
\name{align.distributions}
\alias{align.distributions}
\title{Align joint distributions in table form, to a reference joint distribution}
\usage{
align.distributions(ref.dist.frame, dist.frame.list, include.configsQ = T)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will compute all configuration probabilities as well as logZ
}
\description{
The function will XXXX
}
\details{
The function will XXXX

All joint distributions need to have the same numbers of nodes and configurations
}
