% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{compare_edges}
\alias{compare_edges}
\title{XXXX}
\usage{
compare_edges(model.list, num.nodes = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
XXXX
}
