% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{compare_edges_SAFEISH}
\alias{compare_edges_SAFEISH}
\title{Send in two or more models in the form of a list of crf objects or edge matrices and compare edges}
\usage{
compare_edges_SAFEISH(model.list, num.nodes = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Send in two or more models in the form of a list of crf objects or edge matrices and compare edges
}
\details{
Assumes all models have the same number of nodes and node names are just numbers
}
