% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{complement.at.idx}
\alias{complement.at.idx}
\title{Convenience function to complement a node in a configuration.
Assumes CRF 1,2 states}
\usage{
complement.at.idx(configuration, complement.index)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
