% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_util.R
\name{compute.full.distribution}
\alias{compute.full.distribution}
\title{Compute the joint distribution, with just an input crf object}
\usage{
compute.full.distribution(crf.obj, node.names = NULL, state.names = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will compute all configuration probabilities as well as logZ
}
\description{
The function will XXXX
}
\details{
The function will XXXX

A little shorter version and more generic version of distribution.calc above.
}
