% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_util.R
\name{compute.model.matrix}
\alias{compute.model.matrix}
\title{Utility function to compute model matrix}
\usage{
compute.model.matrix(configs, edges.mat, node.par, edge.par, ff)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Model matrix is the matrix of phi "features". Column
sums of the model matrix should be the sufficient statistics.
}
\details{
The function will XXXX
}

