% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_util.R
\name{conditional.config.energy.grad.test}
\alias{conditional.config.energy.grad.test}
\title{Utility function to compute gradient of negative log pseudo-likelihood}
\usage{
conditional.config.energy.grad.test(config, crf, ff)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Assumes features are 0,1 valued and parameters are numbered.
}
\details{
The function will XXXX
}

