% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{conditional.config.energy.old}
\alias{conditional.config.energy.old}
\title{Older implementation of Energy function for energy of a conditional configuration of states E(Xi | X/Xi).
Uses the feature function formulation.}
\usage{
conditional.config.energy.old(config, condition.element.number,
  adj.node.list, edge.mat, one.lgp, two.lgp, ff, printQ = FALSE)
}
\arguments{
\item{config}{A node configuration (state) vector}

\item{condition.element.number}{i of E(Xi | X/Xi)}

\item{adj.node.list}{XXXX}

\item{two.lgp}{Log node potentials (one-body energies)}

\item{ff}{The feature function}

\item{edges.mat}{Matrix of connected node edges}

\item{two.lgp}{Log edge potentials (two-body energies)}
}
\value{
The function will XX
}
\description{
Assumes log-potentials are in gRbase format
}
\details{
The function will XXXX
}
