% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{conditional.config.energy}
\alias{conditional.config.energy}
\title{Energy function for energy of a conditional configuration of states E(Xi | X/Xi).
Uses the feature function formulation. Eliminates dependency on gRbase format for potentials}
\usage{
conditional.config.energy(
  theta.par = NULL,
  config,
  condition.element.number,
  crf,
  ff,
  printQ = FALSE
)
}
\arguments{
\item{config}{A node configuration (state) vector}

\item{condition.element.number}{i of E(Xi | X/Xi)}

\item{ff}{The feature function}

\item{adj.node.list}{XXXX}

\item{edges.mat}{Matrix of connected node edges}

\item{two.lgp}{Log edge potentials (two-body energies)}
}
\value{
The function will XX
}
\description{
Assumes log-potentials are in gRbase format
}
\details{
The function will XXXX
}
