% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{conditional.config.energy2a}
\alias{conditional.config.energy2a}
\title{Energy function for energy of a conditional configuration of states E(Xi | X/Xi), computed in terms of phi features
This time however pass in a par (theta) vector instead of using the one with the crf object/}
\usage{
conditional.config.energy2a(par, config, phi.config = NULL,
  condition.element.number, crf, ff = NULL)
}
\arguments{
\item{config}{A node configuration (state) vector}

\item{condition.element.number}{i of E(Xi | X/Xi)}

\item{crf}{CRF object}
}
\value{
The function will XX
}
\description{
The function will XXXX
}

