% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{conditional.energy.gradient}
\alias{conditional.energy.gradient}
\title{Gradient of a condtional energy for a node and a configuration}
\usage{
conditional.energy.gradient(
  config,
  condition.element.number,
  crf,
  ff,
  printQ = FALSE
)
}
\arguments{
\item{XXXX}{XXXX}
}
\value{
The function will XX
}
\description{
Gradient of a condtional energy for a node and a configuration
}
\details{
This function compute the vector: \nabla_{\boldsymbol \theta}E(X_i|{\bf X}\slash X_i). The vector is the same length as the
         number of parameters.
}
