% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{config.energy}
\alias{config.energy}
\title{Energy function for a configuration of states}
\usage{
config.energy(config, edges.mat, one.lgp, two.lgp, ff)
}
\arguments{
\item{config}{A node configuration (state) vector}

\item{edges.mat}{Matrix of connected node edges}

\item{two.lgp}{Log node potentials (one-body energies)}

\item{ff}{The feature function}

\item{two.lgp}{Log edge potentials (two-body energies)}
}
\value{
The function will XX
}
\description{
Compute total energy of a configuration. Assumes node/edge energies are in gRbase format
}
\details{
The function will XXXX
}

