% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod.potentials_util.R
\name{config.potential}
\alias{config.potential}
\title{Product of potentials function over a configuration of states}
\usage{
config.potential(config, edges.mat, node.pots, edge.pots)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Product of potentials function
}
\details{
The function will XXXX
}

