% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{copy.crf}
\alias{copy.crf}
\title{(Deep) Copy and return a new independent crf object}
\usage{
copy.crf(crf, plotQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}

