% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{count.row.matches_SAFE}
\alias{count.row.matches_SAFE}
\title{A quick and dirty function to find the frequency of a configuration within a sample of configurations}
\usage{
count.row.matches_SAFE(tab.dat, tab.ref)
}
\arguments{
\item{tab.dat}{The query configuration(s)}

\item{tab.ref}{The reference configurations to search through}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
