% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{count_row_matches}
\alias{count_row_matches}
\title{A quick and dirty function to find the frequency of a configuration within a sample of configurations}
\usage{
count_row_matches(
  tab.query,
  tab.ref,
  drop.query.replicatesQ = T,
  node.names = NULL
)
}
\arguments{
\item{tab.query}{The query configuration(s)}

\item{tab.ref}{The reference configurations to search through}

\item{drop.query.replicatesQ}{If there are repeated entries in tab.query, drop them so their frequencies aren't double counted}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
