% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{crf.dumpout}
\alias{crf.dumpout}
\title{Print out all contents of a crf object}
\usage{
crf.dumpout(crf)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}

