% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_util.R
\name{delta.alpha}
\alias{delta.alpha}
\title{Compute the delta-alpha matrix {\boldsymbol \Delta}_{X_i=1}}
\usage{
delta.alpha(crf, samples, printQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Needed for determination of parameters via logistic regression schemes
}
\details{
The function will XXXX
}

