% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_util.R
\name{distribution.calc}
\alias{distribution.calc}
\title{Compute the joint distribution, assuming a crf object is passed in with gRbase formatted potentials and/or energies}
\usage{
distribution.calc(crf, logZ.calc = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will state probalities in gRbase table form as well as logZ
}
\description{
The function will XXXX
}
\details{
The function will XXXX
}
