% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_util.R
\name{distribution.from.energies}
\alias{distribution.from.energies}
\title{The function will XXXX}
\usage{
distribution.from.energies(state.space, edges.mat, node.energies, edge.energies,
  energy.func, ff)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
The function will XXXX
}
\details{
The function will XXXX
}

