% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{edge.marginal}
\alias{edge.marginal}
\title{Compute an edge marginal}
\usage{
edge.marginal(v.start.node, v.end.node, factor.graph, pots.list, mailbox.list)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Compute a edge marginal
}
\details{
The function will XXXX
}
