% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edges2adj}
\alias{edges2adj}
\title{Form adjacency matrix from edge matrix.
XXXX}
\usage{
edges2adj(edge.mat, plotQ = FALSE)
}
\arguments{
\item{edge.mat}{The XX}
}
\value{
The function will XX
}
\description{
Assumes edges/nodes are numeric.
}
\details{
The function will XXXX
}
