% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{energy.fun1}
\alias{energy.fun1}
\title{Energy function}
\usage{
energy.fun1(config, edges.mat, one.lgp, two.lgp, ff)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Energy function
}
\details{
The function will XXXX
}

