% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{Eone}
\alias{Eone}
\title{One-body (node) energy function}
\usage{
Eone(yA, tA, ff)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
One-body (node) wnergy function
}
\details{
The function will XXXX
}
