% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{Etwo}
\alias{Etwo}
\title{Two-body (edge) energy function}
\usage{
Etwo(yA, yB, wAB, ff)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Two-body (edge) energy function
}
\details{
The function will XXXX
}
