% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_util.R
\name{extract.theta.from.pots}
\alias{extract.theta.from.pots}
\title{Extract parameter vector theta from node and potentials written by trin.mrf}
\usage{
extract.theta.from.pots(crf)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
For testing purposes. train.mrf par and potentials don't always match exactly because of shifting of the pots
}
\details{
The function will XXXX
}

