% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_util.R
\name{feature.means}
\alias{feature.means}
\title{Utility function to compute mean vector of features with supplied theta in a CRF network object.}
\usage{
feature.means(crf, inference.func = infer.exact)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Assumes features are 0,1 valued and parameters are numbered.
}
\details{
The function will XXXX
}

