% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{fit_bayes_logistic}
\alias{fit_bayes_logistic}
\title{Joint distribution from bayes logistic regression fit of parameters}
\usage{
fit_bayes_logistic(
  graph.eq,
  samples,
  iter = 2000,
  thin = 1,
  chains = 4,
  control = NULL
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
