% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{fit_bayes_loglinear2}
\alias{fit_bayes_loglinear2}
\title{Joint distribution from bayes poisson regression fit of parameters, use phi model matrix}
\usage{
fit_bayes_loglinear2(
  graph.eq,
  samples,
  iter = 2000,
  thin = 1,
  chains = 4,
  control = NULL,
  stan.model = NULL
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
