% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{fit_empirical}
\alias{fit_empirical}
\title{Empirical joint distribution}
\usage{
fit_empirical(samples)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Empirical joint distribution
}
\details{
The function will XXXX
}
