% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{fit_logistic}
\alias{fit_logistic}
\title{Joint distribution from logistic regression fit of parameters}
\usage{
fit_logistic(graph.eq, samples)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
