% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{fit_loglinear}
\alias{fit_loglinear}
\title{Joint distribution from Poisson regression fit of parameters}
\usage{
fit_loglinear(graph.eq, samples)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
