% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{fit_mle}
\alias{fit_mle}
\title{Joint distribution from MLE fit of parameters for full MRF loglikelihood}
\usage{
fit_mle(graph.eq, samples, inference.method = infer.exact,
  num.iter = 10, mag.grad.tol = 0.001)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Empirical joint distribution
}
\details{
The function will XXXX
}
