% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{fit_mle_params}
\alias{fit_mle_params}
\title{MLE fit of parameters ONLY for full MRF loglikelihood}
\usage{
fit_mle_params(
  graph.eq,
  samples,
  parameterization.typ = "standard",
  opt.method = "L-BFGS-B",
  inference.method = infer.exact,
  state.nmes = c("1", "2"),
  num.iter = 10,
  mag.grad.tol = 0.001,
  plotQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Use this for bigger models where it is not possible to compute the full joint distribution
}
\details{
The function will XXXX
}
