% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{fit_true}
\alias{fit_true}
\title{Joint distribution from true parameters for MRF}
\usage{
fit_true(true.mrf.modl)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Empirical joint distribution
}
\details{
The function will XXXX
}
