% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.edge_util.R
\name{flatten.marginal.edge.beliefs}
\alias{flatten.marginal.edge.beliefs}
\title{Flatten belief matrices output by marginal.edge.bels to make them stackable}
\usage{
flatten.marginal.edge.beliefs(marginal.edge.beliefs.list)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
