% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{generate_brunner_lizards}
\alias{generate_brunner_lizards}
\title{Generate a little lizard data model from Brunner course for testing}
\usage{
generate_brunner_lizards(
  typ = 1,
  node1.name = NULL,
  node2.name = NULL,
  node3.name = NULL,
  node1.states = NULL,
  node2.states = NULL,
  node3.states = NULL
)
}
\arguments{
\item{XX}{The XX}
}
\value{
A little data set
}
\description{
XXXX
}
\details{
The function will

The function will generate a little lizard data model for testing. Originally from: https://www.utstat.toronto.edu/~brunner/oldclass/312f12/lectures/312f12LoglinearWithR2.pdf
         Generating the data this way saves a little space in testing and allows the user
         to change the names of the nodes and states.
}
