% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{generate_gRbase_lizards}
\alias{generate_gRbase_lizards}
\title{Generate a little lizard data model from the gRbase package for testing}
\usage{
generate_gRbase_lizards()
}
\arguments{
\item{XX}{The XX}
}
\value{
A little data set
}
\description{
XXXX
}
\details{
The function will

The function will generate a little lizard data model for testing. Accessed from the gRbase package and
processed/reconfigured for various testing exercises.
}
