% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{generate_random_model}
\alias{generate_random_model}
\title{Generate a random model using Erdos Renyi game}
\usage{
generate_random_model(
  num.samples = NULL,
  Temp = 1,
  num.nodes = 10,
  p.or.numedges = 0.6,
  type = "gnp",
  seed = NULL,
  plot.sampleQ = F,
  plot.graphQ = T
)
}
\arguments{
\item{num.samples}{Optional number of samples specification}

\item{Temp}{Adjustment parameter to bulk adjust the potentials}

\item{plot.sampleQ}{Optional plot marginal node samples}
}
\value{
A little data set
}
\description{
Generate a random model using Erdos Renyi game
}
\details{
The function will generate a random model using the Erdos Renyi game. Added a "Temp" parameter for user
to adjust the potentials in bulk. Also added sampling from the model.

The function will generate a random model along with a random sample
}
