% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{generate_schmidt_small_model}
\alias{generate_schmidt_small_model}
\title{Generate the original Schmidt small chain model used in Notes}
\usage{
generate_schmidt_small_model(
  num.samples = NULL,
  Temp = 1,
  plot.sampleQ = F,
  plot.graphQ = F
)
}
\arguments{
\item{num.samples}{Optional number of samples specification}

\item{Temp}{Adjustment parameter to bulk adjust the potentials}

\item{plot.sampleQ}{Optional plot marginal node samples}
}
\value{
A little data set
}
\description{
Generate the original Schmidt small chain model used in Notes
}
\details{
The function will generate the original Schmidt small chain model used in Notes. Added a "Temp" parameter for user
to adjust the potentials in bulk. Also added sampling from the model.

The function will generate the original Schmidt small chain model used in Notes along with a random sample
}
