% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_util.R
\name{generate_star_model}
\alias{generate_star_model}
\title{Generate the star model used in their book and the Notes}
\usage{
generate_star_model(
  num.samples = NULL,
  Temp = 1,
  seed = NULL,
  plot.sampleQ = F,
  plot.graphQ = F
)
}
\arguments{
\item{num.samples}{Optional number of samples specification}

\item{Temp}{Adjustment parameter to bulk adjust the potentials}

\item{plot.sampleQ}{Optional plot marginal node samples}
}
\value{
A little data set
}
\description{
Generate the star model used in their book and the Notes
}
\details{
The function will generate the star model used in Notes. Added a "Temp" parameter for user
to adjust the potentials in bulk. Also added sampling from the model.

The function will generate the original Schmidt small chain model used in Notes along with a random sample
}
