% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{get.incoming.messages}
\alias{get.incoming.messages}
\title{Grab incoming messages from the neighboring nodes and put into a list}
\usage{
get.incoming.messages(start.node, end.node, factorgraph, message.list)
}
\arguments{
\item{a.name}{a.name can me the message name or the name of the first node in the messageThe XX}
}
\value{
The function will XX
}
\description{
Grab incoming messages from the neighboring nodes and put into a list
Needed for constructing outgoing messages. NOTE: neighbor node set are all neighbors of
the start.node EXCEPT the end.node.
}
\details{
The function will XXXX
}
