% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_util.R
\name{get.node.idxs}
\alias{get.node.idxs}
\title{Utility function to convert a parameter index to the node/edge indices it is associated with.}
\usage{
get.node.idxs(par.idx, node.par = NULL, edge.par = NULL,
  edge.mat = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Sort of an inverse function of get.par.idx. Intended primarily for checking/debugging.
Assumes features are 0,1 valued and parameters are numbered.
}
\details{
The function will XXXX
}
