% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_util.R
\name{get.par.idx}
\alias{get.par.idx}
\title{Utility function to convert node and edge indices to the
parameter index they are associated with.}
\usage{
get.par.idx(config, i = NULL, j = NULL, node.par = NULL,
  edge.par = NULL, edge.mat = NULL, ff, printQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Assumes features are 0,1 valued and parameters are numbered.
}
\details{
The function will XXXX
}
