% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{get.path}
\alias{get.path}
\title{Get a (shortest) path between two nodes}
\usage{
get.path(a.graphnel.obj, start.node, end.node)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Get a (shortest) path between two nodes
}
\details{
The function will XXXX
}
