% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{get.root.paths}
\alias{get.root.paths}
\title{Get all paths from chosen root node to all leaf nodes}
\usage{
get.root.paths(graph.obj, root.node, serial.schedsQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Get all paths from chosen root node to all leaf nodes
}
\details{
The function will XXXX
}
