% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_configuration_frequencies}
\alias{get_configuration_frequencies}
\title{Get the frequencies of configurations}
\usage{
get_configuration_frequencies(
  sample.mat,
  type = "full",
  empirical.distributionQ = F
)
}
\arguments{
\item{XX}{XXXXXX}
}
\value{
The function will XX
}
\description{
Get the frequencies of configurations. Uses count.row.matches to get partial frequency table
for samples only, but can also get all frequencies (i.e. all the 0 ones too) from the full
contingency table.
}
