% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_util.R
\name{glmnet2graph_info}
\alias{glmnet2graph_info}
\title{Pull info needed for a crf object out of a glmnet fit output}
\usage{
glmnet2graph_info(
  a.glmnet.fit,
  num.nodes,
  crf.obj = NULL,
  lambda = "lambda.1se",
  plotQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Pull info needed for a crf object out of a glmnet fit output
}
\details{
XXXX
}
