% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_util.R
\name{glmnet_logistic_fit2graph_info}
\alias{glmnet_logistic_fit2graph_info}
\title{XXXX}
\usage{
glmnet_logistic_fit2graph_info(
  a.glmnet.logistic.fit,
  num.nodes,
  crf.obj = NULL,
  lambda = "lambda.1se",
  plotQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
XXXX
}
