% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_util.R
\name{grad.loglik}
\alias{grad.loglik}
\title{Utility function to compute gradient of log likelihood}
\usage{
grad.loglik(crf, samples, inference.func = infer.exact)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Assumes features are 0,1 valued and parameters are numbered.
}
\details{
The function will XXXX
}

