% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_util.R
\name{grad.negloglik}
\alias{grad.negloglik}
\title{Utility function to compute gradient of log likelihood}
\usage{
grad.negloglik(crf, nInstances, suffStat, inference.func = infer.exact)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Assumes features are 0,1 valued and parameters are numbered.
}
\details{
The function will XXXX
}
