% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_util.R
\name{grad.neglogpseudolik.config.WRONG}
\alias{grad.neglogpseudolik.config.WRONG}
\title{Utility function to compute gradient of negative log pseudo-likelihood}
\usage{
grad.neglogpseudolik.config.WRONG(config, phi.config = NULL,
  node.conditional.energies = NULL,
  node.complement.conditional.energies = NULL, par = NULL, crf, ff)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Assumes features are 0,1 valued and parameters are numbered.
}
\details{
The function will XXXX
}

