% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_util.R
\name{grad.neglogpseudolik}
\alias{grad.neglogpseudolik}
\title{Gradient of sample negative log pseudo-likelihood}
\usage{
grad.neglogpseudolik(par, crf, samples, ff)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Experiment
}
\details{
The function will XXXX
}

