% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{init.loopy.message.storage}
\alias{init.loopy.message.storage}
\title{Initialize an empty set of list containers to hold messages for loopy belief propagation}
\usage{
init.loopy.message.storage(factor.graph.obj, node.levels = c("up", "dn"))
}
\arguments{
\item{factor.graph.obj}{A factor graph in graphNEL or igraph format.}
}
\value{
The function will XX
}
\description{
Initialize an empty set of list containers to hold messages for loopy belief propagation
No scheduled required to be passed in as all messages are passed at every BP iteration.
}
\details{
The function will XXXX
}
