% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{init.message.storage}
\alias{init.message.storage}
\title{Initialize an empty set of list containers to hold messages}
\usage{
init.message.storage(a.schedule)
}
\arguments{
\item{a.schedule}{A message passing schedule to a root node.}
}
\value{
The function will XX
}
\description{
Initialize an empty set of list containers to hold messages
A NULL message can be interpreted as an id message if necessary??
}
\details{
The function will XXXX
}
