% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_util.R
\name{loglin2crf}
\alias{loglin2crf}
\title{Convert loglin, loglm or dModel (all really loglin) object to a crf object}
\usage{
loglin2crf(log.linear.model.obj, standard.potentialsQ = T, plotQ = F)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Convert loglin, loglm or dModel (all really loglin) object to a crf object
}
\details{
Do glm, glmnet /logistic/poisson model conversion in a different function
}
