% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logsumexp2}
\alias{logsumexp2}
\title{Log sum exp trick. From Brendon Brewer's DNest code:}
\usage{
logsumexp2(logv)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Handy for calculating Z from a vector of log potentials.
A less readable but shorter Log sum exp:
}
\details{
The function will XXXX
}
