% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{make.empty.field}
\alias{make.empty.field}
\title{Instantiate an empty field}
\usage{
make.empty.field(graph.eq = NULL, adj.mat = NULL,
  parameterization.typ = "standard", node.par = NULL, edge.par = NULL,
  plotQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
XXXX
}

