% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{make.f2v.msg}
\alias{make.f2v.msg}
\title{Make factor to variable (node) message}
\usage{
make.f2v.msg(in.v.msgs.list, f.msg, out.v.nme, normalizeQ = F)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Make factor to variable (node) message
\mu_{f\rightarrow X} = \sum_{\backslash X} \Big( f(X) \prod_{Y \in \text{ne}(f)\backslash X} \mu_{Y\rightarrow f} \Big)
}
\details{
The function will XXXX
}
