% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{make.gRbase.beliefs}
\alias{make.gRbase.beliefs}
\title{Decorate fit node and edge marginal beliefs to make compatible with gRbase}
\usage{
make.gRbase.beliefs(inference.obj, node.names, edge.mat, state.nmes = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Decorate fit node and edge marginal beliefs to make compatible with gRbase
}
\details{
The function will XXXX
}
