% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{make.gRbase.potentials}
\alias{make.gRbase.potentials}
\title{Modify initalized mrf-object to make potehtials compatible with gRbase}
\usage{
make.gRbase.potentials(crf, node.names, state.nmes = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Modify initalized mrf-object to make potehtials compatible with gRbase
}
\details{
The function will XXXX
}

