% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_util.R
\name{make.lattice}
\alias{make.lattice}
\title{Quick generate a lattice graph}
\usage{
make.lattice(
  num.rows,
  num.cols,
  node.names.vec = NULL,
  cross.linksQ = T,
  plotQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Quick generate a lattice graph
}
\details{
The function will XXXX
}
