% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentials_util.R
\name{make.par.from.all.potentials}
\alias{make.par.from.all.potentials}
\title{Make a new parameter vector from ALL of the potential values contained in the crf object.
This makes new node/edge pars and returns them. If node/edge pars are contained in the crf object passed in,
they are ignored.}
\usage{
make.par.from.all.potentials(crf)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
The function will XXXX
}
