% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentials_util.R
\name{make.par.from.potentials}
\alias{make.par.from.potentials}
\title{Make a new parameter vector from the potentials contained in the crf object.}
\usage{
make.par.from.potentials(crf)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
update.mrf within train reorders things to the point where
we cannot determine which potentials it utimately re-normalizes. Therefore just
make a new parameter vector.
}
\details{
The function will XXXX
}

