% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentials_util.R
\name{make.pots}
\alias{make.pots}
\title{Make stand alone potentials from input parameter vector.}
\usage{
make.pots(parms, crf, rescaleQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Does same thing as shift.pots but returns potentials instead of putting them into the CRF object.
Eventually this could be combined with shift.pots
}
\details{
The function will XXXX
}

