% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{make.v2f.msg}
\alias{make.v2f.msg}
\title{Make factor to variable (node) message}
\usage{
make.v2f.msg(in.f.msgs.list = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Make variable (node) to factormessage
XXXXXX
}
\details{
The function will XXXX
}
