% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.edge.plot_util.R
\name{marginal.edge.bayes.bels.plot}
\alias{marginal.edge.bayes.bels.plot}
\title{marginal edge empirical probs}
\usage{
marginal.edge.bayes.bels.plot(posterior.edge.belief.info, type = "X1|X2",
  ymax = 2500, edge.empirical.prob.info = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Empirical Pr(X1), Pr(X2), Pr(X1,X2), Pr(X1|X2), Pr(X2|X1) estimates from edge data
The function will XXXX
}
