% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.edge_util.R
\name{marginal.edge.bayes.loglin}
\alias{marginal.edge.bayes.loglin}
\title{marginal edge fit using vanalla baysian poisson regression with rstanarm}
\usage{
marginal.edge.bayes.loglin(edge.samples, prior.sd = NULL,
  prior_intercept.sd = NULL, stan.iter = NULL, prob.level = 0.95,
  printQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXXXXX
The function will XXXX
}
