% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.edge.independence_util.R
\name{marginal.edge.bayes.potentials.indep.indic}
\alias{marginal.edge.bayes.potentials.indep.indic}
\title{marginal edge bayes independence indicators}
\usage{
marginal.edge.bayes.potentials.indep.indic(
  rescaled.posterior.potentials,
  node.names = NULL,
  state.names = NULL,
  prob.level = 0.95,
  printQ = FALSE
)
}
\arguments{
\item{rescaled.posterior.potentials}{Rescaled posterior potentials from marginal.edge.bayes.loglin.}
}
\value{
The function will XX
}
\description{
Outputs statistics on posteriors for tau1, tau2, omega
Handy for checking for independencies between X1 and X2
}
