% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.edge_util.R
\name{marginal.edge.bels.bayes.BROKEN}
\alias{marginal.edge.bels.bayes.BROKEN}
\title{Marginal edge beliefs from bayes posterior sample}
\usage{
marginal.edge.bels.bayes.BROKEN(post.edge.param.samp, node.names = NULL,
  state.names = NULL, printQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Get edge beliefs Pr(X1), Pr(X2), Pr(X1,X2), Pr(X1|X2), Pr(X2|X1) from bayes loglin fit
The function will XXXX
}
