% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.edge_util.R
\name{marginal.edge.bels}
\alias{marginal.edge.bels}
\title{Marginal edge beliefs from parameter point estimates.}
\usage{
marginal.edge.bels(
  edge.mrf.obj,
  node.names = NULL,
  state.names = NULL,
  printQ = FALSE
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Get edge beliefs Pr(X1), Pr(X2), Pr(X1,X2), Pr(X1|X2), Pr(X2|X1) from fit MRF
The function will XXXX
}
