% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.edge_util.R
\name{marginal.edge.emp.pr}
\alias{marginal.edge.emp.pr}
\title{marginal.edge.emp.pr}
\usage{
marginal.edge.emp.pr(edge.samples, printQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Empirical Pr(X1), Pr(X2), Pr(X1,X2), Pr(X1|X2), Pr(X2|X1) estimates from edge data
The function will XXXX
}
