% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.edge_util.R
\name{marginal.edge.loglin}
\alias{marginal.edge.loglin}
\title{marginal edge fit using poisson regression glm}
\usage{
marginal.edge.loglin(edge.samples, conf.level = 0.95, printQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXXXXX
The function will XXXX
}
