% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{message.type}
\alias{message.type}
\title{Get message type depending on name}
\usage{
message.type(a.name)
}
\arguments{
\item{a.name}{a.name can me the message name or the name of the first node in the messageThe XX}
}
\value{
The function will XX
}
\description{
Get message type depending on name
a.name can me the message name or the name of the first node in the message
}
\details{
The function will XXXX
}
