% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{mrf.exact.nll}
\alias{mrf.exact.nll}
\title{Work around for CRF bug evaluating nll for different infer.method's other than on chains}
\usage{
mrf.exact.nll(par, crf, instances, infer.method = infer.exact, ...)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Work around for CRF bug evaluating nll for different infer.method's other than on chains
}
\details{
The function will XXXX
}
