% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_util.R
\name{mrf.glmnet.logistic.fit}
\alias{mrf.glmnet.logistic.fit}
\title{Fit a MRF via cross-validation glmnet LASSO logistic regression}
\usage{
mrf.glmnet.logistic.fit(
  samples,
  crf.obj,
  lambda = "lambda.1se",
  ff,
  infer.func,
  plotQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Fit a MRF via cross-validation glmnet LASSO logistic regression
}
\details{
The function will XXXX
}
