% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{mrf.sample.plot}
\alias{mrf.sample.plot}
\title{Plot a a sample of states node-wise}
\usage{
mrf.sample.plot(samples)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XX
}
\details{
The function will XXXX
}
