% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_util.R
\name{mrf.standard.fit}
\alias{mrf.standard.fit}
\title{Fit a MRF paramiterized in the standard way}
\usage{
mrf.standard.fit(
  samples,
  net.graph.formula,
  num.states,
  mrf.nll.func,
  infer.func
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Fit a MRF paramiterized in the standard way
}
\details{
The function will XXXX
}
