% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{mrf2pwfg2}
\alias{mrf2pwfg2}
\title{Convert Markov Random Field to pairwise (two-body) factor graph:}
\usage{
mrf2pwfg2(graph.obj, plotQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Convert Markov Random Field to pairwise (two-body) factor graph:
}
\details{
This version allows for nodes names to be characters. It just spits out a warning if they are.
}
