% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expts.R
\name{MRF_Stat_expt}
\alias{MRF_Stat_expt}
\title{Experiment}
\usage{
MRF_Stat_expt(crf, instances)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Experiment
}
\details{
The function will XXXX
}

