% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_util.R
\name{neglogpseudolik.config}
\alias{neglogpseudolik.config}
\title{Utility function to compute negative log pseudo-likelihood}
\usage{
neglogpseudolik.config(param = NULL, config, crf, ff,
  cond.en.form = "feature.function")
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Assumes features are 0,1 valued and parameters are numbered.
}
\details{
The function will XXXX
}

