% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_util.R
\name{neglogpseudolik2}
\alias{neglogpseudolik2}
\title{Sample negative log pseudo-likelihood}
\usage{
neglogpseudolik2(
  param,
  crf,
  samples,
  conditional.energy.function.type = "feature.function",
  ff,
  gradQ = FALSE,
  update.crfQ = TRUE
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Experiment
}
\details{
The function will XXXX
}
