% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{nex}
\alias{nex}
\title{Get all neighbors of start.node except end.node}
\usage{
nex(a.graphnel.obj, start.node, end.node)
}
\arguments{
\item{a.name}{a.name can me the message name or the name of the first node in the messageThe XX}
}
\value{
The function will XX
}
\description{
Get all neighbors of start.node except end.node
XXXX
}
\details{
The function will XXXX
}
