% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{node.marginal}
\alias{node.marginal}
\title{Compute a node marginal}
\usage{
node.marginal(v.node.name, factor.graph, mailbox.list)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Compute a node marginal
}
\details{
The function will XXXX
}
