% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nodes2params.list2}
\alias{nodes2params.list2}
\title{Function to find which nodes are associated with which parameters.
Generalization experiment}
\usage{
nodes2params.list2(crf, storeQ = FALSE)
}
\arguments{
\item{crf}{The XX}

\item{storeQ}{Logical, whether or not to store the the node to parameter list in the crf object}
}
\value{
The function will XX
}
\description{
The list produced maps node indices to the parameters they are asociated with
}
\details{
The function will XXXX
}
