% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentials_util.R
\name{par2logpots}
\alias{par2logpots}
\title{Reformat parameter vector to CRF object from node.pot and edge.pot BUT don't
exponentiate, i.e. leave them as logpots (energies).}
\usage{
par2logpots(par, crf)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
The function will XXXX
}
