% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{params2nodes.list}
\alias{params2nodes.list}
\title{Function to find which parameters are associated with which nodes.
XXXX}
\usage{
params2nodes.list(crf, storeQ = FALSE)
}
\arguments{
\item{crf}{The XX}

\item{storeQ}{Logical, whether or not to store the the parameter to node list in the crf object}
}
\value{
The function will XX
}
\description{
The list produced maps parameters indices to the nodes they are asociated with
}
\details{
The function will XXXX
}
