% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.product_util.R
\name{pass.message}
\alias{pass.message}
\title{Pass a message over an edge on a factor graph}
\usage{
pass.message(
  start.node,
  end.node,
  factor.graph,
  pots.list,
  mailboxes.list,
  printQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Pass a message over an edge on a factor graph
}
\details{
The function will XXXX
}
