% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_util.R
\name{phi.component}
\alias{phi.component}
\title{Utility function to compute phi component CONDITIONED for specific node or edge}
\usage{
phi.component(config, i = NULL, j = NULL, node.par = NULL,
  edge.par = NULL, edge.mat = NULL, ff)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Assumes features are 0,1 valued and parameters are numbered.
}
\details{
The function can be used to compute the CONDITINAL phi feature vector used for pseudolikelihood calculations.
}
