% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_util.R
\name{phi.features}
\alias{phi.features}
\title{Utility function to compute unconditional features phi_i(config)}
\usage{
phi.features(config, edges.mat, node.par, edge.par, ff)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Assumes features are 0,1 valued and parameters are numbered.
}
\details{
The function will XXXX
}
