% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{plot_crf}
\alias{plot_crf}
\title{Plot a crf object}
\usage{
plot_crf(crf.obj, type = "g", samples = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XX
}
\details{
The function will XXXX
}
