% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_util.R
\name{process.logistic.fit.JAGS}
\alias{process.logistic.fit.JAGS}
\title{Process JAGS output for logistic fit}
\usage{
process.logistic.fit.JAGS(jags.obj)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
