% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_util.R
\name{process.logistic.fit.stan}
\alias{process.logistic.fit.stan}
\title{Process Stan output for logistic fit}
\usage{
process.logistic.fit.stan(stan.obj, param.vec.name, hpdi.unc = 0.95)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
