% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{reorder_configs}
\alias{reorder_configs}
\title{Spits out permutation to re-order configs in targ with respect to ref}
\usage{
reorder_configs(ref, targ)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
