% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentials_util.R
\name{shift.par}
\alias{shift.par}
\title{Shift node parameter matrices the order update.mrf shifted the node potentials to}
\usage{
shift.par(crf)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
For testing purposes
}
\details{
The function will XXXX
}

