% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentials_util.R
\name{shift.pots}
\alias{shift.pots}
\title{Shift node potentials back to the originally fit parameter vector}
\usage{
shift.pots(crf)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
For testing purposes
}
\details{
The function will XXXX
}
