% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_util.R
\name{sim.field.random}
\alias{sim.field.random}
\title{Simulate an MRF with random unary (node) and pair (edge) potentials.}
\usage{
sim.field.random(adjacentcy.matrix, num.states, num.sims, seed = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Simulate an MRF with random unary (node) and pair (edge) potentials.
}
\details{
Uses standard parameterization w11=w22, w12=w21
}
