% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_util.R
\name{sim.ising.field.random}
\alias{sim.ising.field.random}
\title{Simulate an MRF with random unary (node) and pair (edge) potentials.}
\usage{
sim.ising.field.random(adjacentcy.matrix, parameterization.type = "a",
  num.sims, seed = NULL)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Simulate an MRF with random unary (node) and pair (edge) potentials.
}
\details{
Uses xx standard parameterization
}
