% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_util.R
\name{sim.random.field}
\alias{sim.random.field}
\title{Simulate an random MRF with random unary (node) and pair (edge) potentials.}
\usage{
sim.random.field(
  adj.mat = NULL,
  num.nodes,
  num.edges = NULL,
  prob.edge = 1,
  node.pot.nu = 30,
  node.pot.mean,
  node.pot.sd,
  edge.pot.nu = 30,
  edge.pot.mean,
  edge.pot.sd,
  num.sims,
  seed.node.pots = NULL,
  seed.edge.pots = NULL,
  seed.sample = NULL,
  plotQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Simulate an random MRF with random unary (node) and pair (edge) potentials.
}
\details{
Uses standard parameterization w11=w22, w12=w21. Graph is generated with erdos.renyi.game
in igraph. If an edge appears in the graph, a relatively high to high potential (i.e. non-zero)
should be generated to ensure the edge is potentially strong enough to be found from a random
sample from the graph. However, the user can choose the the general strength and variability of
the edge and node potentials though the edge.pot.mean/sd and node.pot.mean/sd arguments. They
are generated with a t-distribution. The default nu is 30 making the random selections approximately
normal. The number of edges in a randomly generated graph can be chosen by specification or by
probability.
}
