% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UGM_MRF_computeSuffStat.R
\name{UGM_MRF_computeSuffStat}
\alias{UGM_MRF_computeSuffStat}
\title{Port of Schmidt UGM_MRF_computeSuffStat.m}
\usage{
UGM_MRF_computeSuffStat(crf, instances)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
The sufficient statistics seem to be the number
of non-zero energy parameters... See loops in function below
}
\details{
The function will XXXX
}

