% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expts.R
\name{UGM_MRF_makePotentials}
\alias{UGM_MRF_makePotentials}
\title{Experiment: Port of Schmidt UGM_MRF_makePotentials.m}
\usage{
UGM_MRF_makePotentials(w, crf)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Experiment
}
\details{
The function will XXXX
}

