% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{update_CRFutil}
\alias{update_CRFutil}
\title{Easily update the CRFutil library by installing the current version from the github site}
\usage{
update_CRFutil()
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Easily update the CRFutil library by installing the current version from the github site
}
