% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{conditional.config.energy2}
\alias{conditional.config.energy2}
\title{Energy function for energy of a conditional configuration of states E(Xi | X/Xi), computed in terms of phi features}
\usage{
conditional.config.energy2(config, phi.config = NULL,
  condition.element.number, crf, ff = NULL)
}
\arguments{
\item{config}{A node configuration (state) vector}

\item{condition.element.number}{i of E(Xi | X/Xi)}

\item{crf}{CRF object}
}
\value{
The function will XX
}
\description{
Compute
}
\details{
The function will XXXX
}

