% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_util.R
\name{grad.neglogpseudolik.config}
\alias{grad.neglogpseudolik.config}
\title{Utility function to SEPARATELY compute gradient negative log pseudo-likelihood.}
\usage{
grad.neglogpseudolik.config(param = NULL, config, crf, ff,
  cond.en.form = "feature.function")
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Assumes features are 0,1 valued and parameters are numbered.
}
\details{
neglogpseudolik.config() computes the gradient as as well. This is just a pull out for testing and pedegogy.
}

