% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_util.R
\name{KLD}
\alias{KLD}
\title{Compute the Kullback-Leibler distance between two distributions}
\usage{
KLD(px, py, base = exp(1))
}
\arguments{
\item{px}{vector of probability densities p(x)}

\item{py}{vector of probability densities p(y)}
}
\value{
The function will XX
}
\description{
Credit: Shamelessly taken from the excellent LaplacesDemon package
}
\details{
The function will XXXX
}
