% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentials_util.R
\name{make.par.from.potentials}
\alias{make.par.from.potentials}
\title{This version makes a parameter vector from the potentials according to the node/edge pars contained in the
crf object. If the potentials are not scalled according to the codeing in node/edge par, they are re-scaled
before being used.}
\usage{
make.par.from.potentials(crf)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
The function will XXXX
}
