% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentials_util.R
\name{make.pots}
\alias{make.pots}
\title{Make potentials from input parameter vector.}
\usage{
make.pots(parms, crf, rescaleQ = FALSE, replaceQ = FALSE,
  format = "regular", printQ = FALSE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
The function will XXXX
}
