% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.edge.independence_util.R
\name{marginal.edge.bayes.beleifs.indep.indic}
\alias{marginal.edge.bayes.beleifs.indep.indic}
\title{marginal edge bayes independence indicators from beliefs}
\usage{
marginal.edge.bayes.beleifs.indep.indic(rescaled.posterior.potentials,
  node.names = NULL, state.names = NULL, prob.level = 0.95,
  printQ = FALSE)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
Outputs statistics on posteriors for Pr(X1), Pr(X1|X2), Pr(X2), Pr(X2|X1)
Handy for checking for independencies between X1 and X2
}
