% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.edge_util.R
\name{marginal.edge.mrf}
\alias{marginal.edge.mrf}
\title{marginal.edge.mrf}
\usage{
marginal.edge.mrf(edge.samples)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXXXXX
The function will XXXX
}
