% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_util.R
\name{neglogpseudolik}
\alias{neglogpseudolik}
\title{Sample negative log pseudo-likelihood}
\usage{
neglogpseudolik(par, crf, samples,
  conditional.energy.function.type = "feature.function", ff,
  update.crfQ = TRUE)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Experiment
}
\details{
The function will XXXX
}

