% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UGM_MRF_makePotentials.R
\name{UGM_MRF_makePotentials}
\alias{UGM_MRF_makePotentials}
\title{Port of Schmidt UGM_MRF_makePotentials.m}
\usage{
UGM_MRF_makePotentials(w, crf)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Pretty much a direct port of Schmidt's UGM_MRF_makePotentials.m
}
\details{
The function will XXXX
}

